Complete get/set Attribute and toJSONString

+++ ProductTree/Storage.java
@@ -1,6 +1,6 @@
 abstract class Storage extends Product {
-    protected int capacity; // Unit: GB
-    protected int readSpeed; // Unit: MB/s
-    protected int writeSpeed; // Unit: MB/s
+    protected Long capacity; // Unit: GB
+    protected Long readSpeed; // Unit: MB/s
+    protected Long writeSpeed; // Unit: MB/s
     protected String slot;
 }
